Extract shared AppState from CoAP module to state.rs
Move AppState struct, APP_STATE mutex, and with_app_state() accessor
from coap.rs into state.rs so both CoAP and the upcoming Matter
handler can share the same application state. Add init_app_state()
helper to replace direct mutex access during registration.
